Fix for the initialisation of NaN and Inf crashing on RO 5.
On RISC OS 5 (and 3.7) it appears that the initialisation of the
NaN and Inf variables was crashing. It's unclear why that is, but
to avoid the problem the variables are now initialised with their
constant values which should ensure that the tool starts, even
if the actual use of such values might be bad.

Version number incremented to 1.40.