Fixes for the XMLLint tool not writing content properly.
The XMLLint tool wasn't writing the content properly to the output,
or reading the catalogs correctly. The output to the screen had been
broken in the prior ports, and hadn't been made to work. It has now
been updated so that we use a new function that operates on file
handles, rather than system handles.

The catalog handling had become broken in an earlier version of the
library, after the last version of the port. The handling of the
catalogs is not functional if there is no stat support - it will
always believe that a file is present and never go to the catalog
code. We now have a stub which identifies the file as existing,
which gets around the problem that stat does not exist.