Add some simple tests that the built xmllint works.
The built XMLLint needs to be tested as part of the CI process. This
is obviously important, but more so when the released tools have been
non-functional for two releases. Really I should listen to my own
words.

These tests really do only exercise a few of the functions of the tool,
but they're at least enough to show that the tool is working.