Added --native command line switch.
We can now specify native RISC OS filenames on the command line. This
is pretty much required for any RISC OS port, so it is surprising that
I didn't even offer the option. The output handling is now much more
reliable - the output filenames had gone through multiple transformations
which meant that they were effectively RISC OS format when the input
filenames were in URI format.

A couple of tests are performed on the tool to check that the native
switch works properly.